Refactor pretalx.common.settings.utils
